Move NewLinodeClient and ErrorHelper to the new helper package
