+++ builder/linode/step_create_image.go
@@ -6,6 +6,7 @@ import (
 	"github.com/hashicorp/packer-plugin-sdk/multistep"
 	packersdk "github.com/hashicorp/packer-plugin-sdk/packer"
 	"github.com/linode/linodego"
+	"github.com/linode/packer-plugin-linode/helper"
 )
 
 type stepCreateImage struct {
@@ -19,7 +20,7 @@ func (s *stepCreateImage) Run(ctx context.Context, state multistep.StateBag) mul
 	instance := state.Get("instance").(*linodego.Instance)
 
 	handleError := func(prefix string, err error) multistep.StepAction {
-		return errorHelper(state, ui, prefix, err)
+		return helper.ErrorHelper(state, ui, prefix, err)
 	}
 
 	creationPoller, err := s.client.NewEventPoller(
