+++ builder/linode/step_shutdown_linode.go
@@ -6,6 +6,7 @@ import (
 	"github.com/hashicorp/packer-plugin-sdk/multistep"
 	packersdk "github.com/hashicorp/packer-plugin-sdk/packer"
 	"github.com/linode/linodego"
+	"github.com/linode/packer-plugin-linode/helper"
 )
 
 type stepShutdownLinode struct {
@@ -18,7 +19,7 @@ func (s *stepShutdownLinode) Run(ctx context.Context, state multistep.StateBag) 
 	instance := state.Get("instance").(*linodego.Instance)
 
 	handleError := func(prefix string, err error) multistep.StepAction {
-		return errorHelper(state, ui, prefix, err)
+		return helper.ErrorHelper(state, ui, prefix, err)
 	}
 
 	ui.Say("Shutting down Linode...")
